requirements.txt
click==8.1.8
colorama==0.4.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
fsspec==2025.3.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.1.0
networkx==2.8.8
numpy==1.26.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvrtc-cu11==11.8.89
opencv-python==4.9.0.80
pillow==11.1.0
python-dotenv==1.0.1
sympy==1.13.1
